LayoutTests/media/video-dom-loopend.html

<video src=content/test.mp4 loopcount=2 autoplay></video>
<script src=video-test.js></script>
<script>
video.loopEnd=0.5;
test("video.loopEnd==0.5");
test("video.getAttribute('loopend') == '0.5s'");
setTimeout(function () { test("video.currentTime<=0.5"); }, 2000);
waitForEventTestAndEnd('timeupdate', "video.currentTime.toFixed(1) == 0 && video.currentLoop==1");
</script>
