LayoutTests/media/video-dom-loopend-expected.txt

TEST(video.loopEnd==0.5) OK
TEST(video.getAttribute('loopend') == '0.5s') OK
EVENT(timeupdate) TEST(video.currentTime.toFixed(1) == 0 && video.currentLoop==1) OK
END OF TEST
